+++ Themaproject/src/java/com/oncloud6/atd/rights/GroupsRightsRestoreServlet.java
@@ -39,7 +39,15 @@ public class GroupsRightsRestoreServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        MySQLConnection DBConnection = new MySQLConnection();
+        MySQLConnection DBConnection = new MySQLConnection();HttpSession session = request.getSession(true);
+        RequestDispatcher rd = null;
+        RightsControl.initRequest(request, response);
+        int userId = Integer.parseInt(session.getAttribute("groupID").toString());
+        if(!RightsControl.checkBoolean("groups_rights_restore", "true", userId)) {
+            rd = request.getRequestDispatcher("error/403error.jsp");
+            rd.forward(request, response);
+            return;
+        }
         try {
             Connection connect = DBConnection.getConnection();
             PreparedStatement preparedStatement = connect.prepareStatement("DELETE FROM atd.rechten_groepen WHERE rechten_id = ? AND groepen_id = ?");
@@ -54,9 +62,6 @@ public class GroupsRightsRestoreServlet extends HttpServlet {
             //niet vergeten om alles te sluiten :)
             preparedStatement.close();
             connect.close();
-
-            RequestDispatcher rd = null;
-            HttpSession session = request.getSession(true);
 
             response.sendRedirect("groupsedit?id=" + gid);
